requirements.txt
# Core dependencies
openai>=1.0.0              # For OpenAI and DeepSeek LLM calls
python-dotenv>=1.0.0       # For loading environment variables
pyyaml>=6.0                # For YAML parsing
flask>=2.3.0               # For API server
flask-cors>=4.0.0          # For CORS support

# Optional dependencies (uncomment if needed)
# google-generativeai>=0.3.0  # For Google Gemini support
# duckduckgo-search>=3.8.0   # For DuckDuckGo search (no API key required)
# requests>=2.28.0           # For web search APIs (Serper, Tavily, Brave, Bocha)